refactor : 패키지 분리 상세화
