+++ src/main/java/janggi/movement/crash/CrashOrBridgeAndPrey.java
@@ -1,9 +1,9 @@
-package janggi.point.crash;
+package janggi.movement.crash;
 
 import janggi.game.Team;
+import janggi.movement.route.Hurdles;
 import janggi.piece.Movable;
 import janggi.piece.Po;
-import janggi.point.Hurdles;
 import janggi.point.Point;
 import java.util.List;
