+++ src/main/java/janggi/movement/route/Route.java
@@ -1,7 +1,9 @@
-package janggi.point;
+package janggi.movement.route;
 
 import janggi.piece.Movable;
-import janggi.point.crash.Crashes;
+import janggi.point.Point;
+import janggi.movement.crash.Crashes;
+import janggi.movement.direction.Direction;
 import java.util.ArrayList;
 import java.util.List;
